[#29] Turn `Command` into a GADT
Problem: we want to have a single `runCommand` function that is able to
run any command. This means we need to be able to know what each command
should return when executed.

Solution: Turn `Command` into a GADT, and use a phantom type variable to
specify what each command should return.