[#29] Report parsing failures for field keys and tags
Problem: In the CLI, if we fail to parse a `FieldKey` or `EntryTag`, we
want to be able to tell the user the exact reason why parsing failed.
Right now, the smart constructors for these types only return a
`Nothing` when they fail, so it's impossible to know why they failed.

Solution: change the smart constructors to return `Either Text a`, with
an explanation for the failure.