Basis for a Vault specific library embedded in coffer
Problem: A library is needed to communicate with Vault and to marshall
values to and from Vault's JSON representations.

Solution: Add two modules, one representing an abstract concept of a
`SecretEngine` and one implementing a specific secrets engine, namely
`Kv`, for Vault's kv2. The idea is that multiple secret engines may
eventually be supported. For example kv1 or cubbyhole. The generalized
marshalling framework `EntryConvertible` wasn't utilized as that's
intended for more storage and UI usecases. Eventually it may be worth
switching to it.

Signed-off-by: Magic_RB <magic_rb@redalder.org>