Change the return type of `readSecret` and `listSecrets` to `Maybe a`
Problem: An `Entry` not being present in a given backend is not really
an error, but rather valid state. Therefore it shouldn't be grouped
together with real errors like `ConnentionError`.

Solution: Handle the case when Vault returns a 404 in a specal way and
if it happens don't throw an error but return `Nothing`.

Signed-off-by: Magic_RB <magic_rb@redalder.org>