Add a polysemy effect implementing operations with secrets
Problem: A developer utilizing coffer should not need to care about for
which particular backend they're developing for.

Solution: The `Vault` effect serves this purpose, the developer can
program in terms of the generic effect functions `readSecret`,
`listSecret`. What actual backend is used, is decided later, by the
means of the interpreter used to interpert the afformentioned effect.

Signed-off-by: Magic_RB <magic_rb@redalder.org>