Implement a generic backend and config parsing system
Problem: The user should be able to configure multiple different
backends, for two corporate instances of Vault and one personal pass
backend. The users configuration should be compeletely transparent to
frontend developers, therefore there must be a function `Config ->
BackendName -> BackendInterpreter`.

Solution: Create a `Config` parser, which according to a simple matching
function returns a `PackedBackend`, which can be passed around without
knowing the actual type of the backend inside, then as neccessary it can
bu unpacked into an `a :: Backend` which again erases the type. This can
then be used to get the interpreter of the `BackendEffect` for that
particular backend. No special handling is necessary when implementing a
new backend.

Signed-off-by: Magic_RB <magic_rb@redalder.org>