[Chore] Rename `code_style.md`
Problem: In `.github/pull_request_template.md` we have a reference to
`docs/code-style.md`, but this reference is broken. The file actually
uses an underscore instead of an hyphen.

Solution: We tend to use an underscore in the name of `code-style.md` in
other projects, so we should rename the file here.