[#26] Fix `validate cabal files` build step
Problem: The `validate cabal files` build step hasn't been working ever
since we added it. We can see this message in the logs:

```
./scripts/validate-cabal-files.sh: line 39: diff: command not found
```

The reason why the build step did not fail (thus, it was a false
negative) is because we temporarily use `set +e` to mask all non-zero
exit codes.

Solution: add the missing `diffTools` to the nix shell.