[#22] Forbid copying entry or directory to itself
Problem: coffer copy -f a a was allowed. It is unnecessary operation,
so user should be informed of this. Renaming entry to itself
was also allowed in force mode, which should be forbidden.

Solution: Check paths for equality before performing the operation.
Change test that expected renaming entry to itself to work.

[#22] Forbid copying entry or directory to itself

Problem: coffer copy -f a a was allowed. It is unnecessary operation,
so user should be informed of this. Renaming entry to itself
was also allowed in force mode, which should be forbidden.

Solution: Check paths for equality before performing the operation.
Change test that expected renaming entry to itself to work.