[#83] Add `coffer` commands handlers
Problem: when a command fails, the Web API returns a 200 status code with
an error message in the body. We should return an appropriate
status code depending on the error with an error message in the body.
Each type of error should also be associated with a numeric code.
This will allow the frontend to know exactly what went wrong and act appropriately.

Solution: created a bunch of `coffer` command handlers which take result
and throw appropriate status code. Each command error is associated with
a numeric code according to `web-api-error-codes.md`.