[#106] Backend-agnostic `listDirectoryContents`
Problem: at this moment the `listDirectoryContents` function returns `[Text]`
and in `getEntryOrDir` we rely on the fact that `vault` adds a '/'
to the directory. But other backends might not behave like this.

Solution: created new data `DirectoryContents` which stores 2 lists:
1. list with directories
2. list with entries
and changed return type from `Maybe [Text]` to `Maybe DirectoryContents`
in `listDirectoryContents` function.