[#120] Install the coffer executable before running the bats tests
Problem: before running the bats tests, the user has to remember to
first install the coffer executable. This means if they forget to do
that step, the tests will be run against a stale version of coffer,
which might produce false positives/negatives.

Solution: install the coffer executable in the `make bats` target before
running the tests.